requirements.txt
-i https://pypi.org/simple
aiohttp==3.6.2
appdirs==1.4.4
async-timeout==3.0.1; python_full_version >= '3.5.3'
attrs==19.3.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
certifi==2020.6.20
chardet==3.0.4
idna==2.10; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
inflection==0.5.0
jmespath==0.10.0
multidict==4.7.6; python_version >= '3.5'
orjson==3.2.2
pyee==7.0.2
pyhumps==1.6.1
pyppeteer==0.2.2
requests==2.24.0
termcolor==1.1.0
tqdm==4.48.0; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
urllib3==1.25.9; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
uvloop==0.14.0
websockets==8.1; python_full_version >= '3.6.1'
yarl==1.4.2; python_version >= '3.5'
